+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 [<img height='36' style='border:0px;height:36px;' src='https://az743702.vo.msecnd.net/cdn/kofi2.png?v=2' border='0' alt='Buy Me a Coffee at ko-fi.com'/>](https://ko-fi.com/X8X41IE4T)
 
+## 2.0.8
+
 ## 2.0.7
 * VSCode:
   * Support online-docs-only options in version selector